refactor:#21 Sidebar 컴포넌트 분리
